pyproject.toml
[tool.poetry]
name = "hw-27"
version = "0.1.0"
description = ""
authors = ["Specially4 <kotenkovsn@gmail.com>"]
readme = "README.md"
packages = [{include = "hw_27"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
psycopg2 = "^2.9.3"
Pillow = "^9.2.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
